Restructure the codebase
