+++ configuration_test.go
@@ -1,8 +1,14 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
+	"github.com/alicebob/miniredis"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
 	"os"
+	"rediseen/types"
 	"strings"
 	"testing"
 )
@@ -392,4 +398,922 @@ func Test_dbCheck_expose_all(t *testing.T) {
 			t.Error("something is wrong with dbCheck()")
 		}
 	}
+}
+
+func compareAndShout(t *testing.T, expected interface{}, actual interface{}) {
+	if expected != actual {
+		t.Error("Expecting\n", expected, "\ngot\n", actual)
+	}
+}
+
+func Test_service_wrong_usage(t *testing.T) {
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	expectedCode := 400
+	expectedError := "Usage: /db, /db/key, /db/key/index, or /db/key/field"
+	casesToTest := []string{"/0/", "/0/key:1/", "/0/key:1/1/", "/0/key:1/1/1", "/0/key:1/1/1/", "/0/key:1/1/1/1"}
+	var res *http.Response
+
+	for _, suffix := range casesToTest {
+		res, _ = http.Get(s.URL + suffix)
+
+		if res.StatusCode != expectedCode {
+			t.Error("Expecting\n", expectedCode, "\ngot\n", res.StatusCode)
+		}
+
+		resultStr, _ := ioutil.ReadAll(res.Body)
+		res.Body.Close()
+
+		var result types.ErrorType
+		json.Unmarshal([]byte(resultStr), &result)
+
+		compareAndShout(t, expectedError, result.Error)
+	}
+}
+
+func Test_service_non_integer_db_provided(t *testing.T) {
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/a/key")
+
+	expectedCode := 400
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	expectedError := "Provide an integer for DB"
+	compareAndShout(t, expectedError, result.Error)
+}
+
+func Test_service_redis_conn_refused(t *testing.T) {
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1")
+
+	expectedCode := 500
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	expectedError := "connection refused"
+	if !strings.Contains(result.Error, expectedError) {
+		t.Error("Expecting to contain \n", expectedError, "\ngot\n", result.Error)
+	}
+}
+
+func Test_service_non_existent_key(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1")
+
+	expectedCode := 404
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	expectedError := "Key provided does not exist."
+	compareAndShout(t, expectedError, result.Error)
+}
+
+// Check listing-keys feature
+// Taking keys which are NOT exposed into consideration as well (they should NOT be counted or returned)
+func Test_service_list_keys_by_db_1(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	testSlice1 := []string{"hi", "world", "test", "done"}
+	for i, v := range testSlice1 {
+		mr.Set(fmt.Sprintf("key:%v", i), v)
+	}
+
+	testSlice2 := []string{"r", "a", "n", "d", "o", "m"}
+	for i, v := range testSlice2 {
+		mr.HSet(fmt.Sprintf("key:%v", i+len(testSlice1)), v, v)
+	}
+
+	testSlice3 := []string{"list", "type"}
+	for i, v := range testSlice3 {
+		mr.Lpush(fmt.Sprintf("key:%v", i+len(testSlice1)+len(testSlice2)), v)
+	}
+
+	testSlice4 := []string{"no", "access", "key"}
+	for i, v := range testSlice4 {
+		mr.HSet(fmt.Sprintf("no_access_key:%v", i+len(testSlice1)+len(testSlice2)+len(testSlice3)), v, v)
+	}
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.KeyListType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, len(testSlice1)+len(testSlice2)+len(testSlice3), len(result.Keys))
+	compareAndShout(t, len(testSlice1)+len(testSlice2)+len(testSlice3), result.Count)
+	if result.Total > 1000 {
+		t.Error("Listing keys function should return <= 1000 keys")
+	}
+}
+
+// Check listing-keys feature
+// Check the situation where more than 1000 keys are exposed
+// `Total` should be the total number, BUT only up to 1000 keys should be returned.
+// `Count` should be up to 1000 as well
+func Test_service_list_keys_by_db_2(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	n := 2000
+	for i := 0; i < 2000; i++ {
+		mr.Set(fmt.Sprintf("key:%v", i), string(i))
+	}
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.KeyListType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, 1000, result.Count)
+	compareAndShout(t, n, result.Total)
+}
+
+func Test_service_string_type(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("key:1", "hi")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ResponseType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "string", result.ValueType)
+	compareAndShout(t, "hi", result.Value)
+}
+
+func Test_service_string_check_by_index(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("key:1", "Developer")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1/0")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ResponseType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "string", result.ValueType)
+	compareAndShout(t, "D", result.Value)
+
+	res, _ = http.Get(s.URL + "/0/key:1/4")
+
+	expectedCode = 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ = ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "string", result.ValueType)
+	compareAndShout(t, "l", result.Value)
+
+	res, _ = http.Get(s.URL + "/0/`key:1`/5")
+
+	expectedCode = 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ = ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "string", result.ValueType)
+	compareAndShout(t, "o", result.Value)
+
+	res, _ = http.Get(s.URL + "/0/`key:1`/`6`")
+
+	expectedCode = 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ = ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "string", result.ValueType)
+	compareAndShout(t, "p", result.Value)
+}
+
+func Test_service_string_check_by_index_wrong_index(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("key:1", "Developer")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1/x")
+
+	compareAndShout(t, 400, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	expectedError := strWrongTypeForIndexField
+	compareAndShout(t, expectedError, result.Error)
+}
+
+func Test_service_string_type_with_slash_in_key(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("key:/1", "hi")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/`key:/1`")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ResponseType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "string", result.ValueType)
+	compareAndShout(t, "hi", result.Value)
+}
+
+func Test_service_string_type_with_slash_and_backtick_in_key(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("key:`/1", "hi")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/`key:`/1`")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ResponseType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "string", result.ValueType)
+	compareAndShout(t, "hi", result.Value)
+}
+
+// Validate if listing-keys feature returns 200 if the DB in client request IS exposed
+func Test_service_list_keys_for_db_with_access(t *testing.T) {
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	// env var set for the test is REDISEEN_DB_EXPOSED=0-5
+	for db := 0; db <= 5; db++ {
+		res, _ := http.Get(s.URL + fmt.Sprintf("/%v", db))
+
+		expectedCode := 200
+		compareAndShout(t, expectedCode, res.StatusCode)
+
+		resultStr, _ := ioutil.ReadAll(res.Body)
+		res.Body.Close()
+
+		var result types.KeyListType
+		json.Unmarshal([]byte(resultStr), &result)
+
+		compareAndShout(t, 0, result.Count)
+		compareAndShout(t, 0, result.Total)
+	}
+}
+
+// Validate if listing-keys feature returns 403 if the DB in client request is NOT exposed
+func Test_service_list_keys_for_db_without_access(t *testing.T) {
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	// env var set for the test is REDISEEN_DB_EXPOSED=0-5
+	for db := 6; db <= 100; db++ {
+		res, _ := http.Get(s.URL + fmt.Sprintf("/%v", db))
+
+		expectedCode := 403
+		compareAndShout(t, expectedCode, res.StatusCode)
+
+		resultStr, _ := ioutil.ReadAll(res.Body)
+		res.Body.Close()
+
+		var result types.ErrorType
+		json.Unmarshal([]byte(resultStr), &result)
+
+		expectedError := fmt.Sprintf("DB %v is not exposed", db)
+		compareAndShout(t, expectedError, result.Error)
+	}
+}
+
+func Test_service_string_type_db_no_access(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	// env var set for the test is REDISEEN_DB_EXPOSED=0-5
+	for _, db := range []int{6, 10, 100} {
+		res, _ := http.Get(s.URL + fmt.Sprintf("/%v/key:1", db))
+
+		expectedCode := 403
+		compareAndShout(t, expectedCode, res.StatusCode)
+
+		resultStr, _ := ioutil.ReadAll(res.Body)
+		res.Body.Close()
+
+		var result types.ErrorType
+		json.Unmarshal([]byte(resultStr), &result)
+
+		expectedError := fmt.Sprintf("DB %v is not exposed", db)
+		compareAndShout(t, expectedError, result.Error)
+	}
+}
+
+func Test_service_string_type_key_no_access(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	//env var set for the test is REDISEEN_KEY_PATTERN_EXPOSED=^key:[.]*
+	mr.Set("id:1", "hi")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/id:1")
+
+	expectedCode := 403
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	expectedError := "Key pattern is forbidden from access"
+	compareAndShout(t, expectedError, result.Error)
+}
+
+func Test_service_list_key(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Lpush("key:1", "hello")
+	mr.Lpush("key:1", "world")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"list","value":["world","hello"]}`
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_list_key_check_by_index(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Lpush("key:1", "hello")
+	mr.Lpush("key:1", "world")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1/0")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"list","value":"world"}`
+	compareAndShout(t, expectedResult, string(result))
+
+	res, _ = http.Get(s.URL + "/0/key:1/1")
+
+	expectedCode = 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ = ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult = `{"type":"list","value":"hello"}`
+	compareAndShout(t, expectedResult, string(result))
+
+	// Check wrong type for index
+	res, _ = http.Get(s.URL + "/0/key:1/a")
+
+	expectedCode = 400
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ = ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult = fmt.Sprintf(`{"error":"%s"}`, strWrongTypeForIndexField)
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_set(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.SetAdd("key:1", "hello")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"set","value":["hello"]}`
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_set_check_by_index(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.SetAdd("key:1", "hello")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1/hello")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"set","value":true}`
+	compareAndShout(t, expectedResult, string(result))
+
+	res, _ = http.Get(s.URL + "/0/key:1/world")
+
+	expectedCode = 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ = ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult = `{"type":"set","value":false}`
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_hash(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.HSet("key:1", "role", "developer")
+	mr.HSet("key:1", "id", "1")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"hash","value":{"id":"1","role":"developer"}}`
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_hash_check_by_index(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.HSet("key:1", "role", "developer")
+	mr.HSet("key:1", "id", "1")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:1/role")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"hash","value":"developer"}`
+	compareAndShout(t, expectedResult, string(result))
+
+	res, _ = http.Get(s.URL + "/0/key:1/id")
+
+	expectedCode = 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ = ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult = `{"type":"hash","value":"1"}`
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_zset(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.ZAdd("key:set", 100, "developer")
+	mr.ZAdd("key:set", 0, "bluffer")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:set")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"zset","value":["bluffer","developer"]}`
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_zset_check_by_field(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.ZAdd("key:set", 200, "Mr.X")
+	mr.ZAdd("key:set", 100, "developer")
+	mr.ZAdd("key:set", 0, "bluffer")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	res, _ := http.Get(s.URL + "/0/key:set/developer")
+
+	expectedCode := 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	result, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	expectedResult := `{"type":"zset","value":1}`
+	compareAndShout(t, expectedResult, string(result))
+}
+
+func Test_service_delete_not_allowed(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("key:1", "hello")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	client := &http.Client{}
+	req, _ := http.NewRequest("DELETE", s.URL+"/0/key:1", nil)
+	res, _ := client.Do(req)
+
+	expectedCode := 405
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	expectedError := "Method DELETE is not allowed"
+	compareAndShout(t, expectedError, result.Error)
+}
+
+// Request method will happen first, so whatever method other than GET should always get rejected
+// Checks like key pattern check should not even happen
+func Test_service_delete_not_allowed_no_access(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("id:1", "hello")
+
+	originalRedisURI := os.Getenv("REDISEEN_REDIS_URI")
+	os.Setenv("REDISEEN_REDIS_URI", fmt.Sprintf("redis://:@%s", mr.Addr()))
+	defer os.Setenv("REDISEEN_REDIS_URI", originalRedisURI)
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	client := &http.Client{}
+	req, _ := http.NewRequest("DELETE", s.URL+"/0/id:1", nil)
+	res, _ := client.Do(req)
+
+	expectedCode := 405
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	expectedError := "Method DELETE is not allowed"
+	compareAndShout(t, expectedError, result.Error)
+}
+
+func Test_api_key_authentication(t *testing.T) {
+
+	mr, _ := miniredis.Run()
+	defer mr.Close()
+
+	mr.Set("key:1", "hello")
+
+	os.Setenv("REDISEEN_API_KEY", "nopass")
+	defer os.Setenv("REDISEEN_REDIS_URI", "")
+
+	var config configuration
+	config.loadFromEnv()
+	s := httptest.NewServer(http.Handler(&config))
+	defer s.Close()
+
+	// case-1: no API Key is provided in request header
+	// (to FAIL)
+	res, _ := http.Get(s.URL + "/0")
+
+	expectedCode := 401
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	resultStr, _ := ioutil.ReadAll(res.Body)
+	res.Body.Close()
+
+	var result types.ErrorType
+	json.Unmarshal([]byte(resultStr), &result)
+
+	compareAndShout(t, "unauthorized", result.Error)
+
+	// case-2: correct API Key is provided in request header
+	// (to SUCCEED)
+	client := &http.Client{}
+	req, _ := http.NewRequest("GET", s.URL+"/0", nil)
+	req.Header.Add("X-API-KEY", "nopass")
+	res, _ = client.Do(req)
+
+	expectedCode = 200
+	compareAndShout(t, expectedCode, res.StatusCode)
+
+	// case-3: wrong API Key is provided in request header
+	// (to FAIL)
+	client = &http.Client{}
+	req, _ = http.NewRequest("GET", s.URL+"/0", nil)
+	req.Header.Add("X-API-KEY", "wrongkey")
+	res, _ = client.Do(req)
+
+	expectedCode = 401
+	compareAndShout(t, expectedCode, res.StatusCode)
 }
