+++ rediseen_test.go
@@ -70,6 +70,9 @@ func Test_Main(t *testing.T) {
 	// First element "" is a placeholder for executable
 	//ref: https://stackoverflow.com/a/48674736
 
+	os.Setenv("REDISEEN_REDIS_URI", "redis://:@localhost:6400")
+	defer os.Unsetenv("REDISEEN_REDIS_URI")
+
 	// command "rediseen"
 	os.Args = []string{""}
 	main()
@@ -95,6 +98,9 @@ func Test_Main_invalid_config(t *testing.T) {
 }
 
 func Test_Main_start_command(t *testing.T) {
+	os.Setenv("REDISEEN_REDIS_URI", "redis://:@localhost:6400")
+	defer os.Unsetenv("REDISEEN_REDIS_URI")
+
 	os.Args = []string{"", "-d", "start"}
 	main()
 }
